Remove product-related UI components from the showcase app
- Delete `ProductCard`, `ProductDetailClient`, and `ProductsGrid` components.
- Remove `CartClient` and `CartIcon` components along with associated cart functionality.
- Clean up unused imports, contexts, and schemas related to cart and products.
