+++ apps/showcase/app/[locale]/about/page.tsx
@@ -1,4 +1,5 @@
 import { getTranslations } from 'next-intl/server';
+import { PageHeader } from '@/components/sections/PageHeader';
 
 export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }) {
   const { locale } = await params;
@@ -26,14 +27,7 @@ export default async function AboutPage({ params }: { params: Promise<{ locale: 
   return (
     <div className="py-16 sm:py-24">
       <div className="mx-auto max-w-4xl px-4 sm:px-6 lg:px-8">
-        <div className="text-center mb-12" dir={isRTL ? 'rtl' : 'ltr'}>
-          <h1 className="text-4xl font-bold tracking-tight text-primary sm:text-5xl">
-            {t('title')}
-          </h1>
-          <p className="mt-4 text-lg text-secondary">
-            {t('subtitle')}
-          </p>
-        </div>
+        <PageHeader title={t('title')} subtitle={t('subtitle')} isRTL={isRTL} />
 
         <div className="prose prose-lg max-w-none" dir={isRTL ? 'rtl' : 'ltr'}>
           <section className="mb-12">
