+++ apps/showcase/app/[locale]/contact/page.tsx
@@ -1,140 +1,20 @@
 'use client';
 
-import { useState } from 'react';
 import { useTranslations } from 'next-intl';
 import { useLocale } from 'next-intl';
-import { socialLinks } from '@/config/socialLinks';
+import { PageHeader } from '@/components/sections/PageHeader';
+import { ContactForm } from '@/components/forms/ContactForm';
 
 export default function ContactPage() {
   const t = useTranslations('contact');
   const locale = useLocale();
   const isRTL = locale === 'ar';
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    message: ''
-  });
-  const [status, setStatus] = useState<'idle' | 'sending' | 'success' | 'error'>('idle');
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setStatus('sending');
-    
-    // Simulate form submission
-    setTimeout(() => {
-      setStatus('success');
-      setFormData({ name: '', email: '', message: '' });
-      setTimeout(() => setStatus('idle'), 3000);
-    }, 1000);
-  };
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    setFormData(prev => ({
-      ...prev,
-      [e.target.name]: e.target.value
-    }));
-  };
 
   return (
     <div className="py-16 sm:py-24">
       <div className="mx-auto max-w-2xl px-4 sm:px-6 lg:px-8">
-        <div className="text-center mb-12" dir={isRTL ? 'rtl' : 'ltr'}>
-          <h1 className="text-4xl font-bold tracking-tight text-primary sm:text-5xl">
-            {t('title')}
-          </h1>
-          <p className="mt-4 text-lg text-secondary">
-            {t('subtitle')}
-          </p>
-        </div>
-
-        <div className="rounded-lg bg-white p-8 shadow-lg border border-border" dir={isRTL ? 'rtl' : 'ltr'}>
-          <form onSubmit={handleSubmit} className="space-y-6">
-            <div>
-              <label htmlFor="name" className="block text-sm font-medium text-primary mb-2">
-                {t('name')}
-              </label>
-              <input
-                type="text"
-                id="name"
-                name="name"
-                required
-                value={formData.name}
-                onChange={handleChange}
-                className="w-full rounded-md border border-border px-4 py-2 text-primary shadow-sm focus:border-secondary focus:ring-secondary bg-background"
-              />
-            </div>
-
-            <div>
-              <label htmlFor="email" className="block text-sm font-medium text-primary mb-2">
-                {t('email')}
-              </label>
-              <input
-                type="email"
-                id="email"
-                name="email"
-                required
-                value={formData.email}
-                onChange={handleChange}
-                className="w-full rounded-md border border-border px-4 py-2 text-primary shadow-sm focus:border-secondary focus:ring-secondary bg-background"
-              />
-            </div>
-
-            <div>
-              <label htmlFor="message" className="block text-sm font-medium text-primary mb-2">
-                {t('message')}
-              </label>
-              <textarea
-                id="message"
-                name="message"
-                required
-                rows={6}
-                value={formData.message}
-                onChange={handleChange}
-                className="w-full rounded-md border border-border px-4 py-2 text-primary shadow-sm focus:border-secondary focus:ring-secondary bg-background"
-              />
-            </div>
-
-            <button
-              type="submit"
-              disabled={status === 'sending'}
-              className="w-full rounded-md bg-primary px-6 py-3 text-base font-semibold text-white shadow-sm transition-colors hover:bg-secondary focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {status === 'sending' ? t('sending') : t('send')}
-            </button>
-
-            {status === 'success' && (
-              <div className="rounded-md bg-accent p-4 text-primary">
-                {t('success')}
-              </div>
-            )}
-
-            {status === 'error' && (
-              <div className="rounded-md bg-accent p-4 text-secondary">
-                {t('error')}
-              </div>
-            )}
-          </form>
-
-          <div className="mt-8 border-t border-border pt-8">
-            <h3 className="text-lg font-semibold text-primary mb-4">
-              {t('socialMedia')}
-            </h3>
-            <div className="flex gap-4">
-              {socialLinks.map((link) => (
-                <a
-                  key={link.name}
-                  href={link.url}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="text-secondary hover:text-primary transition-colors"
-                  aria-label={link.ariaLabel}
-                >
-                  {link.icon}
-                </a>
-              ))}
-            </div>
-          </div>
-        </div>
+        <PageHeader title={t('title')} subtitle={t('subtitle')} isRTL={isRTL} />
+        <ContactForm isRTL={isRTL} />
       </div>
     </div>
   );
