+++ apps/showcase/app/[locale]/page.tsx
@@ -1,6 +1,7 @@
 import { getTranslations } from 'next-intl/server';
 import { products } from '@/data/products';
-import ProductsGrid from '@/components/ProductsGrid/ProductsGrid';
+import ProductsGrid from '@/components/products/ProductsGrid';
+import { Hero } from '@/components/sections/Hero';
 
 export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }) {
   const { locale } = await params;
@@ -28,18 +29,7 @@ export default async function HomePage({ params }: { params: Promise<{ locale: s
   return (
     <div className="flex flex-col">
       {/* Header Section */}
-      <section className="relative overflow-hidden bg-gradient-to-br from-accent/30 to-background py-12 sm:py-16">
-        <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
-          <div className="mx-auto max-w-2xl text-center" dir={isRTL ? 'rtl' : 'ltr'}>
-            <h1 className="text-3xl font-bold tracking-tight text-primary sm:text-4xl">
-              {t('title')}
-            </h1>
-            <p className="mt-4 text-base text-secondary">
-              {t('subtitle')}
-            </p>
-          </div>
-        </div>
-      </section>
+      <Hero title={t('title')} subtitle={t('subtitle')} isRTL={isRTL} />
 
       {/* Products Grid */}
       <section className="py-8 sm:py-12">
