+++ apps/showcase/app/[locale]/products/page.tsx
@@ -1,6 +1,7 @@
 import { getTranslations } from 'next-intl/server';
 import { products } from '@/data/products';
-import ProductsGrid from '@/components/ProductsGrid/ProductsGrid';
+import ProductsGrid from '@/components/products/ProductsGrid';
+import { PageHeader } from '@/components/sections/PageHeader';
 
 export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }) {
   const { locale } = await params;
@@ -28,14 +29,7 @@ export default async function ProductsPage({ params }: { params: Promise<{ local
   return (
     <div className="py-16 sm:py-24">
       <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
-        <div className="mx-auto max-w-2xl text-center mb-12" dir={isRTL ? 'rtl' : 'ltr'}>
-          <h1 className="text-4xl font-bold tracking-tight text-primary sm:text-5xl">
-            {t('title')}
-          </h1>
-          <p className="mt-4 text-lg text-secondary">
-            {t('subtitle')}
-          </p>
-        </div>
+        <PageHeader title={t('title')} subtitle={t('subtitle')} isRTL={isRTL} />
 
         {products.length > 0 ? (
           <ProductsGrid products={products} />
