+++ apps/showcase/components/ui/ImageCarousel.tsx
@@ -20,7 +20,8 @@ export default function ImageCarousel({
   isRTL,
 }: ImageCarouselProps) {
   const [selectedIndex, setSelectedIndex] = useState(0);
-  
+  const [hoveredIndex, setHoveredIndex] = useState<number | null>(null);
+
   // Combine all media: main image first, then video if exists, then additional images
   const allMedia = [
     { type: 'image' as const, src: mainImage },
@@ -28,27 +29,32 @@ export default function ImageCarousel({
     ...images.map((img) => ({ type: 'image' as const, src: img })),
   ];
 
-  const selectedMedia = allMedia[selectedIndex];
+  // Use hovered index if available, otherwise use selected index
+  const displayIndex = hoveredIndex !== null ? hoveredIndex : selectedIndex;
+  const selectedMedia = allMedia[displayIndex];
 
   const handleThumbnailClick = (index: number) => {
     setSelectedIndex(index);
+    setHoveredIndex(null); // Clear hover when clicking
   };
 
   const handlePrevious = () => {
     setSelectedIndex((prev) => (prev === 0 ? allMedia.length - 1 : prev - 1));
+    setHoveredIndex(null); // Clear hover when navigating
   };
 
   const handleNext = () => {
     setSelectedIndex((prev) => (prev === allMedia.length - 1 ? 0 : prev + 1));
+    setHoveredIndex(null); // Clear hover when navigating
   };
 
   return (
-    <div className="flex flex-col gap-4" dir={isRTL ? 'rtl' : 'ltr'}>
+    <div className="flex gap-4" dir={isRTL ? 'rtl' : 'ltr'}>
       {/* Main Image/Video Display */}
-      <div className="relative aspect-square overflow-hidden rounded-lg bg-accent/20 border border-border group">
+      <div className="relative flex-1 aspect-square overflow-hidden rounded-lg bg-accent/20 border border-border group">
         <AnimatePresence mode="wait">
           <motion.div
-            key={selectedIndex}
+            key={displayIndex}
             initial={{ opacity: 0 }}
             animate={{ opacity: 1 }}
             exit={{ opacity: 0 }}
@@ -71,7 +77,7 @@ export default function ImageCarousel({
                 width={800}
                 height={800}
                 className="h-full w-full object-cover transition-transform duration-700 group-hover:scale-105"
-                priority={selectedIndex === 0}
+                priority={displayIndex === 0}
               />
             )}
           </motion.div>
@@ -82,11 +88,11 @@ export default function ImageCarousel({
           <>
             <button
               onClick={handlePrevious}
-              className="absolute left-4 top-1/2 -translate-y-1/2 z-10 rounded-full bg-white/90 p-2 shadow-lg opacity-0 transition-opacity group-hover:opacity-100 hover:bg-white focus:opacity-100 focus:outline-none focus:ring-2 focus:ring-primary"
+              className={`absolute top-1/2 -translate-y-1/2 z-10 rounded-full bg-white/90 p-2 shadow-lg opacity-0 transition-opacity group-hover:opacity-100 hover:bg-white focus:opacity-100 focus:outline-none focus:ring-2 focus:ring-primary ${isRTL ? 'right-2' : 'left-2'}`}
               aria-label={isRTL ? "الصورة التالية" : "Image précédente"}
             >
               <svg
-                className={`h-5 w-5 text-primary ${isRTL ? 'rotate-180' : ''}`}
+                className={`h-3.5 w-3.5 text-primary ${isRTL ? 'rotate-180' : ''}`}
                 fill="none"
                 stroke="currentColor"
                 viewBox="0 0 24 24"
@@ -101,11 +107,11 @@ export default function ImageCarousel({
             </button>
             <button
               onClick={handleNext}
-              className="absolute right-4 top-1/2 -translate-y-1/2 z-10 rounded-full bg-white/90 p-2 shadow-lg opacity-0 transition-opacity group-hover:opacity-100 hover:bg-white focus:opacity-100 focus:outline-none focus:ring-2 focus:ring-primary"
+              className={`absolute top-1/2 -translate-y-1/2 z-10 rounded-full bg-white/90 p-2 shadow-lg opacity-0 transition-opacity group-hover:opacity-100 hover:bg-white focus:opacity-100 focus:outline-none focus:ring-2 focus:ring-primary ${isRTL ? 'left-2' : 'right-2'}`}
               aria-label={isRTL ? "الصورة السابقة" : "Image suivante"}
             >
               <svg
-                className={`h-5 w-5 text-primary ${isRTL ? 'rotate-180' : ''}`}
+                className={`h-3.5 w-3.5 text-primary ${isRTL ? 'rotate-180' : ''}`}
                 fill="none"
                 stroke="currentColor"
                 viewBox="0 0 24 24"
@@ -134,14 +140,16 @@ export default function ImageCarousel({
         )}
       </div>
 
-      {/* Thumbnail Navigation */}
+      {/* Thumbnail Navigation - Vertical on Right */}
       {allMedia.length > 1 && (
-        <div className="flex gap-2 overflow-x-auto pb-2 scrollbar-hide">
+        <div className="flex flex-col gap-2 overflow-y-auto pb-2 scrollbar-hide">
           {allMedia.map((media, index) => (
             <button
               key={index}
               onClick={() => handleThumbnailClick(index)}
-              className={`relative flex-shrink-0 aspect-square w-20 overflow-hidden rounded-lg border-2 transition-all ${
+              onMouseEnter={() => setHoveredIndex(index)}
+              onMouseLeave={() => setHoveredIndex(null)}
+              className={`relative flex-shrink-0 aspect-square w-12 overflow-hidden rounded-lg border-2 transition-all ${
                 selectedIndex === index
                   ? 'border-primary scale-105'
                   : 'border-border hover:border-secondary'
@@ -151,14 +159,14 @@ export default function ImageCarousel({
                 <div className="relative h-full w-full bg-accent/20">
                   <div className="absolute inset-0 flex items-center justify-center">
                     <svg
-                      className="h-8 w-8 text-primary"
+                      className="h-4 w-4 text-primary"
                       fill="currentColor"
                       viewBox="0 0 24 24"
                     >
                       <path d="M8 5v14l11-7z" />
                     </svg>
                   </div>
-                  <div className={`absolute top-1 ${isRTL ? 'left-1' : 'right-1'} rounded bg-black/70 px-1.5 py-0.5 text-xs font-medium text-white`}>
+                  <div className={`absolute top-0.5 ${isRTL ? 'left-0.5' : 'right-0.5'} rounded bg-black/70 px-1 py-0.5 text-[10px] font-medium text-white`}>
                     {isRTL ? 'فيديو' : 'Vidéo'}
                   </div>
                 </div>
@@ -166,8 +174,8 @@ export default function ImageCarousel({
                 <Image
                   src={media.src || '/images/placeholder.jpg'}
                   alt={`${productName} - Vue ${index + 1}`}
-                  width={80}
-                  height={80}
+                  width={48}
+                  height={48}
                   className="h-full w-full object-cover"
                 />
               )}
